Phase B2: extract document-writer helpers into src/config_writes/
